pyproject.toml
[project]
name = "bar"
version = "0.1.0"
description = "code & model for arxiv paper \"Autoregressive Image Generation with Masked Bit Modeling\""
readme = "README.md"
requires-python = "~=3.13.0"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.1",
    "einops>=0.8.2",
    "gdown>=5.2.1",
    "huggingface-hub>=1.4.1",
    "iopath>=0.1.10",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "open-clip-torch>=3.2.0",
    "pillow>=12.1.1",
    "scipy>=1.17.0",
    "termcolor>=3.3.0",
    "timm>=1.0.24",
    "torch-fidelity>=0.3.0",
    "torchinfo>=1.8.0",
    "tqdm>=4.67.3",
    "transformers>=5.1.0",
    "wandb>=0.25.0",
    "tensorboard>=2.20.0",
    "webdataset>=1.0.2",
    "xformers>=0.0.32.post2",
    "torch~=2.10.0",
    "torchvision",
    "flash-attn",
    "tensorboard>=2.20.0",
    "pycocoevalcap>=1.2",
    "python-dotenv>=1.2.2",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
flash-attn = { url = "https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.16/flash_attn-2.8.3%2Bcu130torch2.10-cp313-cp313-linux_x86_64.whl" }

[[tool.uv.index]]
url = "https://mirrors.sustech.edu.cn/pypi/web/simple"
default = true

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true
